Add code to let ingesters report their name, version, and UUID to the indexer.
The IngestConnection type has gained an IdentifyIngester function which
is used to send the aforementioned info to the indexer. The muxer calls
this function whenever it establishes a connection to one of its indexers.

The muxer is informed of name/version/UUID by setting the fields
IngesterName, IngesterVersion, and IngesterUUID in the UniformMuxerConfig
or MuxerConfig structs. Leaving these blank is not a problem.

If the indexer has not reported a sufficiently high version, the ingest
library will silently ignore the request to identify the ingester,
since it is purely an informational call for users on the indexer.