Rework how ingesters log.
Instead of having a bunch of logic for the ingest muxer to send logs to the
gravwell tag, instead just provide convenience functions which write to
the Logger which was passed in the config. In turn, that logger can set
the ingest muxer as a *relay*, which writes the log messages to the gravwell
tag. This has the side effect of getting *all* log messages into the gravwell
tag, provided you don't fill up the (generous) buffer before finally connecting
to an indexer.