Add Max-Buffer and Trim-Whitespace options to regex listener.
Max-Buffer tells the listener how many bytes it is allowed to store
while looking for a regex match before it should give up and just dump
a partial entry.

Trim-Whitespace is a boolean; if set to true, entries will have whitespace
trimmed from both ends before shipping.

This commit also splits out the basic operation of the regex matching
into a function which is easier to call from tests.