client: Do a sync in the Login function.
Previously, we told users to call Login, then call Sync, which fetches
information about the user. This tends to be a bit fraught in practice,
because users reasonably assume that after calling NewClient and Login,
you should be able to use the client -- but some calls *will* fail,
because the client doesn't know the user's UID. This change just
calls syncNoLock as the last step of Login.