Fixing a potential race in the file follower where a new directory is created and a file immediately written to it.  There is a gap of time where the fsnotify watcher for the file is still being wired up and the file lands, as a result the kernel does not notify us of the file creation.
Scenario, you mv a folder full of files into a watched directory.
File Follower gets the creation event on the folder, files are moved
in almost instantly.  File follower checks out the directory and wires
up the watcher; but files made it in before you could watch it, so
you never get the fsnotify event.

I was only ever able to make this race happen on a virtualized Redhat7
box.  Something changed about fsnotify on modern kernels so that I still
get events in the above situation on a modern system.