Attempt to make the Splunk job more intelligent
We previously just tried to find chunks of N entries or less. Now we
also try to stay under M megabytes (800, because we've seen limits of
1GB for a single query in some situations). Tracks average size of
entries seen so far in order to estimate how big the chunk will be,
because tstats doesn't keep track of entry sizes, just counts. Because
of this, we now start with a 1 minute chunk so we can get spun up
quick and estimate entry sizes sooner than later, in order to expand
our window if possible.

Also prints out more useful information if the job runs into
unexpected results; often as not, it's Splunk spitting back XML
because you've exceeded some quota.