First pass at fields to enable write access.
We want to do query library first, which means we get to start with
Things, the most complicated part. This adds information about writers
to the Thing type.

It also adds a field to the WireSearchLibrary which is used to
communicate to the client what the current user is allowed to modify
on any given item; only owners and admins can change sharing fields,
but anyone in the new WriteAccess lists can change the other stuff.