Add MinVersion field to muxer configs.
Muxer will refuse to connect to any indexer reporting an API version
lower than that specified in the config, thus if you intend to use
ditto you might say:

		igCfg := ingest.UniformMuxerConfig{
			[...]
			MinVersion:         ingest.MINIMUM_DITTO_VERSION,
			[...]
		}
		igst, _ = ingest.NewUniformMuxer(igCfg)

An insufficient API version is treated as non-fatal; the muxer will
attempt to reconnect to that indexer periodically, on the assumption
that it will eventually be updated.

As long as one indexer meets the minimum version requirement, you will
be able to use the muxer as usual. If not, the muxer will never go hot.